+++ R/moves_calc.R
@@ -10,7 +10,8 @@ utils::globalVariables(c("moves", "switches", ".SD"))
 #' time without being credited with having moved at least once.
 #' @param gap a number, of days,  that represents the largest gap between an exit date 
 #' and the next enrollment date that can occur without indicating the student 
-#' moved to a third school not contained within the data set. The default value is \code{14}.
+#' moved to a third school not contained within the data set. The default value is 
+#' \code{14}.
 #' @param sid  a character that indicates the name of the student id attribute 
 #' in \code{df}. The default value is \code{sid}.
 #' @param schid a character that indicates the name of the school id attribute 
@@ -19,9 +20,12 @@ utils::globalVariables(c("moves", "switches", ".SD"))
 #' attribute in \code{df}. The default value is \code{enroll_date}.
 #' @param exit_date a character that indicates the name of the student id 
 #' attribute in \code{df}. The default value is \code{exit_date}.
+#' @details \code{enrollby} and \code{exitby} are specified automatically if not 
+#' defined. They are assigned to the default dates of -09-15 and -06-01 of the min 
+#' and max year respectively.
 #' @author Jason P. Becker
 #' @import data.table
-#' @return a data.table
+#' @return a data.frame
 #' @export
 #'
 #' @examples